+++ components/Input/index.tsx
@@ -48,13 +48,13 @@ export const withInputStyleProps =
 
     let inputShallot: InputShallot = {
       Container: {
-        borderRadius: getRadius(radius),
+        borderRadius: getRadius(String(radius)),
         backgroundColor:
           color === 'Default'
             ? getGlobal('backgroundColor')
-            : getColor(color, 50),
+            : getColor(String(color), 50),
         borderColor:
-          color === 'Default' ? getGlobal('borderColor') : getColor(color, 700),
+          color === 'Default' ? getGlobal('borderColor') : getColor(String(color), 700),
         display: 'flex',
         borderWidth: 1,
         borderStyle: 'solid',
@@ -68,13 +68,13 @@ export const withInputStyleProps =
       Input: {
         display: 'flex',
         flexGrow: 1,
-        fontFamily: getFontFamily(fontFamily),
+        fontFamily: getFontFamily(String(fontFamily)),
         color:
           color === 'Default'
             ? getGlobal('foregroundColor')
-            : getColor(color, 900),
-        fontSize: getFontSize(fontSize),
-        letterSpacing: getLetterSpacing(letterSpacing),
+            : getColor(String(color), 900),
+        fontSize: getFontSize(String(fontSize)),
+        letterSpacing: getLetterSpacing(String(letterSpacing)),
 
         ...(uppercase && { textTransform: uppercase ? 'uppercase' : 'none' }),
 
@@ -89,7 +89,7 @@ export const withInputStyleProps =
           borderColor:
             color === 'Default'
               ? getGlobal('foregroundColor')
-              : getColor(color, 900),
+              : getColor(String(color), 900),
         },
         Input: {
           outline: 'none',