+++ core/mixins/src/getFontFamily/index.ts
@@ -11,8 +11,8 @@ import { valueNotFoundError } from '../utils'
 export const getFontFamily =
   <T extends DefaultTheme = DefaultTheme>(
     rawKey: keyof T['tokens']['fontFamilies'],
-  ): MixinFunction<string> =>
-  ({ theme }): CSS.Properties['fontFamily'] => {
+  ): MixinFunction<string | undefined> =>
+  ({ theme }): string | undefined => {
     const key = rawKey as keyof (typeof theme)['tokens']['fontFamilies']
     const value = theme?.tokens?.fontFamilies?.[key]
     if (!value) console.warn(valueNotFoundError('fontFamilies', String(key)))