+++ core/mixins/src/getFontSize/index.ts
@@ -11,8 +11,8 @@ import { valueNotFoundError } from '../utils'
 export const getFontSize =
   <T extends DefaultTheme = DefaultTheme>(
     rawKey: keyof T['tokens']['fontSizes'],
-  ): MixinFunction<string | number> =>
-  ({ theme }): number | undefined => {
+  ): MixinFunction<string | number | undefined> =>
+  ({ theme }): number | string | undefined => {
     const key = rawKey as keyof (typeof theme)['tokens']['fontSizes']
     const value = theme?.tokens?.fontSizes?.[key]
     if (value === undefined) {