+++ core/mixins/src/getShadow/index.ts
@@ -13,9 +13,9 @@ export const getShadow =
   <T extends DefaultTheme = DefaultTheme>(
     rawKey: keyof T['tokens']['shadows'],
   ): MixinFunction<string> =>
-  ({ theme }): CSS.Properties['boxShadow'] => {
+  ({ theme }): string => {
     const key = rawKey as keyof (typeof theme)['tokens']['shadows']
     const value = theme?.tokens?.shadows?.[key]
     if (!value) console.warn(valueNotFoundError('shadows', String(key)))
-    return value
+    return value || ''
   }