+++ core/mixins/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shallot-ui/core-mixins
 
+## 3.1.0
+
+### Minor Changes
+
+- 7b000ab: feat: removes references to csstype
+
 ## 3.0.0
 
 ### Minor Changes