+++ packages/email/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @shallot-ui/email
 
+## 4.3.3
+
+### Patch Changes
+
+- Updated dependencies [7b000ab]
+  - @shallot-ui/core-mixins@3.1.0
+  - @shallot-ui/core-props@4.0.0
+  - @shallot-ui/core-utils@4.0.0
+  - @shallot-ui/platform-react-email@4.0.0
+
 ## 4.3.2
 
 ### Patch Changes